Update on Part 4 with corrections
I re-added the second graphs for the male data because in parts 1,2,3,4 of question 4. I realized that, due to our ongoing commit issues in the previous depository, they hadn’t been saved properly. While fixing that, I also noticed I had made a mistake in parts 1 and 2 of question 4. For question 1, I had originally included the region hues, but that was incorrect since the question doesn’t ask for a third feature. The third feature should only be included in question 2, where the region variable serves as that extra feature. Therefore, i removed the population size feature, because if not we would have had 4 features in question 2 when it only asks for 3.