Completed part 6
Completed part c and d by printing our the exact variables that we wanted to compare in order to do our analysis. Since for part 6c) we want to look at region, internet use and emissions per capita. i printed it so we can clearly see it without having to refer back to the excel dataset.